Move utils for tests
